requirements.txt
langchain
langchain-community
huggingface_hub
sentence-transformers
faiss-cpu
streamlit
python-dotenv
pandas
transformers
torch
